st/nine: Move Set/Get/Free-PrivateData to IUnknown.
Moving the methods Set/Get/Free-PrivateData to IUnknown allows
to reuse the same code for Volume9 objects.

The duplicated code in Volume9 is removed.
